requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cors-headers==3.12.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
envparse==0.2.0
greenlet==1.1.2
iniconfig==1.1.1
kombu==5.2.4
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.1.0
wrapt==1.14.1
